+++ packages/api-client/src/v2/features/app/app-events.ts
@@ -110,6 +110,16 @@ export function initializeAppEventHandlers({
           },
         }),
       },
+      'operation:create:operation': {
+        onAfterExecute: (payload) => {
+          onAfterExampleCreation({
+            path: payload.path,
+            method: payload.method,
+            exampleKey: 'default',
+            documentName: payload.documentName,
+          })
+        },
+      },
       'operation:upsert:parameter': {
         onAfterExecute: (payload) => onAfterExampleCreation(payload.meta),
       },