+++ packages/api-reference/src/components/Content/Operations/TraversedEntry.vue
@@ -50,6 +50,7 @@ const {
     | 'orderRequiredPropertiesFirst'
     | 'orderSchemaPropertiesBy'
     | 'showOperationId'
+    | 'hideModels'
   >
   /** Currently selected server for the document */
   selectedServer: ServerObject | null