+++ integrations/java/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/java-integration
 
+## 0.6.22
+
 ## 0.6.21
 
 ## 0.6.20