+++ packages/api-client-react/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/api-client-react
 
+## 2.0.1
+
 ## 2.0.0
 
 ### Major Changes