+++ packages/openapi-to-markdown/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/openapi-to-markdown
 
+## 0.5.3
+
 ## 0.5.2
 
 ### Patch Changes