+++ packages/pre-post-request-scripts/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @scalar/scripts
 
+## 0.4.1
+
 ## 0.4.0
 
 ### Minor Changes